+++ appliance/infrastructures/Symfony/Recipe/Step/Job/CallNewJob.php
@@ -29,9 +29,13 @@ use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Teknoo\East\Common\View\ParametersBag;
 use Teknoo\East\Foundation\Manager\ManagerInterface;
+use Teknoo\East\Paas\Contracts\Message\MessageInterface;
+use Teknoo\East\Paas\Contracts\Security\EncryptionInterface;
+use Teknoo\Recipe\Promise\Promise;
 use Teknoo\Space\Contracts\Recipe\Step\Job\CallNewJobInterface;
 use Teknoo\Space\Object\DTO\NewJob;
 use Teknoo\Space\Object\DTO\SpaceProject;
+use Throwable;
 
 /**
  * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
@@ -43,6 +47,7 @@ class CallNewJob implements CallNewJobInterface
 {
     public function __construct(
         private MessageBusInterface $messageBus,
+        private ?EncryptionInterface $encryption,
     ) {
     }
 
@@ -52,11 +57,28 @@ class CallNewJob implements CallNewJobInterface
         SpaceProject $project,
         ParametersBag $parametersBag,
     ): CallNewJobInterface {
-        $this->messageBus->dispatch(
-            new Envelope(
-                $newJob->export()
-            )
-        );
+        $dispatching = function (NewJob $newJob): void {
+            $this->messageBus->dispatch(
+                new Envelope(
+                    $newJob->export()
+                )
+            );
+        };
+
+        if (null === $this->encryption) {
+            $dispatching($newJob);
+        } else {
+            /** @var Promise<MessageInterface, mixed, mixed> $promise */
+            $promise = new Promise(
+                onSuccess: $dispatching,
+                onFail: fn (Throwable $error) => throw $error,
+            );
+
+            $this->encryption->encrypt(
+                data: $newJob,
+                promise: $promise,
+            );
+        }
 
         $parametersBag->set('newJobId', $newJob->newJobId);
         $parametersBag->set('projectId', $project->getId());